chore: 회원, 신고 엔티티 seq 데이터 타입 변경

+++ matzipback/src/main/java/com/matzip/matzipback/report/command/domain/aggregate/Penalty.java
@@ -16,12 +16,12 @@ public class Penalty {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int penaltySeq;
+    private long penaltySeq;
 
     /*@ManyToOne
     @JoinColumn(name = "penalty_user_seq", referencedColumnName = "userSeq")
     private Users penaltyUserSeq;*/
-    private int penaltyUserSeq;
+    private long penaltyUserSeq;
 
     @CreatedDate
     private LocalDateTime penalty_start_date;
