+++ matzipback/src/main/java/com/matzip/matzipback/users/command/domain/aggregate/Messages.java
@@ -10,6 +10,6 @@ public class Messages {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int messageSeq;
+    private long messageSeq;
 
 }
